057. Correction error with validation

+++ Controllers/ProductController.cs
@@ -75,59 +75,68 @@ namespace Rocky.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Upsert(ProductVM productVM, Product product)
         {
-            var files = HttpContext.Request.Form.Files;
-            string webRootPath = _webHostEnvironment.WebRootPath;
-
-            if (productVM.Product.Id == 0)
+            if (ModelState.IsValid)
             {
-                // Creating
-                string upload = webRootPath + WC.ImagePath;
-                string fileName = Guid.NewGuid().ToString();
-                string extension = Path.GetExtension(files[0].FileName);
 
-                using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
+                var files = HttpContext.Request.Form.Files;
+                string webRootPath = _webHostEnvironment.WebRootPath;
+
+                if (productVM.Product.Id == 0)
                 {
-                    files[0].CopyTo(fileStream);
-                }
-
-                productVM.Product.Image = fileName + extension;
-
-                _db.Product.Add(productVM.Product);
-            }
-            else 
-            {
-               // updating
-               var objFromDb = _db.Product.AsNoTracking().FirstOrDefault( x =>x.Id  == productVM.Product.Id);
-
-                if (files.Count > 0)
-                {
+                    // Creating
                     string upload = webRootPath + WC.ImagePath;
                     string fileName = Guid.NewGuid().ToString();
                     string extension = Path.GetExtension(files[0].FileName);
-
-                    var oldFile = Path.Combine(upload, objFromDb.Image);
-
-                    if (System.IO.File.Exists(oldFile))
-                    {
-                        System.IO.File.Delete(oldFile);
-                    }
 
                     using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
                     {
                         files[0].CopyTo(fileStream);
                     }
-                    productVM.Product.Image = fileName + extension;
-                }
 
+                    productVM.Product.Image = fileName + extension;
+
+                    _db.Product.Add(productVM.Product);
+                }
                 else
                 {
-                    productVM.Product.Image = objFromDb.Image;
+                    // updating
+                    var objFromDb = _db.Product.AsNoTracking().FirstOrDefault(x => x.Id == productVM.Product.Id);
+
+                    if (files.Count > 0)
+                    {
+                        string upload = webRootPath + WC.ImagePath;
+                        string fileName = Guid.NewGuid().ToString();
+                        string extension = Path.GetExtension(files[0].FileName);
+
+                        var oldFile = Path.Combine(upload, objFromDb.Image);
+
+                        if (System.IO.File.Exists(oldFile))
+                        {
+                            System.IO.File.Delete(oldFile);
+                        }
+
+                        using (var fileStream = new FileStream(Path.Combine(upload, fileName + extension), FileMode.Create))
+                        {
+                            files[0].CopyTo(fileStream);
+                        }
+                        productVM.Product.Image = fileName + extension;
+                    }
+
+                    else
+                    {
+                        productVM.Product.Image = objFromDb.Image;
+                    }
+                    _db.Product.Update(productVM.Product);
                 }
-                _db.Product.Update(productVM.Product);
+                _db.SaveChanges();
+                return RedirectToAction("Index");
             }
-            _db.SaveChanges();
-            return RedirectToAction("Index");
-            
+            productVM.CategorySelectList = _db.Category.Select(i => new SelectListItem
+            {
+                Text = i.Name,
+                Value = i.Id.ToString()
+            });
+             return View(productVM);
         }
     }
 }
